hotfix: emptyPrayCardList 처리
