+++ src/components/todayPray/TodayPrayCardList.tsx
@@ -12,6 +12,13 @@ import { KakaoShareButton } from "../share/KakaoShareBtn";
 import MyMemberBtn from "../member/MyMemberBtn";
 import { PrayTypeDatas } from "@/Enums/prayType";
 import OpenContentDrawerBtn from "../share/OpenContentDrawerBtn";
+import {
+  Drawer,
+  DrawerContent,
+  DrawerDescription,
+  DrawerHeader,
+  DrawerTitle,
+} from "@/components/ui/drawer";
 
 interface PrayCardListProps {
   currentUserId: string;
@@ -27,6 +34,14 @@ const TodayPrayCardList: React.FC<PrayCardListProps> = ({
     (state) => state.fetchGroupPrayCardList
   );
 
+  const isPrayToday = useBaseStore((state) => state.isPrayToday);
+  const isOpenTodayPrayDrawer = useBaseStore(
+    (state) => state.isOpenTodayPrayDrawer
+  );
+  const setIsOpenTodayPrayDrawer = useBaseStore(
+    (state) => state.setIsOpenTodayPrayDrawer
+  );
+
   const prayCardCarouselApi = useBaseStore(
     (state) => state.prayCardCarouselApi
   );
@@ -36,24 +51,7 @@ const TodayPrayCardList: React.FC<PrayCardListProps> = ({
 
   const startDt = getISOTodayDate(-6);
   const endDt = getISOTodayDate(1);
-
-  const completedItem = (
-    <div className="flex flex-col gap-4 justify-center items-center min-h-[80vh] max-h-[80vh] pb-10">
-      <img
-        src={PrayTypeDatas["pray"].img}
-        alt={PrayTypeDatas["pray"].emoji}
-        className="w-16 h-16 opacity-100"
-      />
-      <h1 className="font-bold text-xl">오늘의 기도 완료!</h1>
-      <h3 className="text-gray-600">내일도 기도해 주실 거죠? 🤗</h3>
-      <div className="text-gray-400 text-center">
-        <h1>당신을 위해 기도한</h1>
-        <h1>친구들을 확인해 보아요</h1>
-      </div>
-      <MyMemberBtn />
-      <OpenContentDrawerBtn />
-    </div>
-  );
+  const todayDt = getISOTodayDate();
 
   useEffect(() => {
     fetchGroupPrayCardList(groupId, currentUserId, startDt, endDt);
@@ -82,36 +80,54 @@ const TodayPrayCardList: React.FC<PrayCardListProps> = ({
       </div>
     );
 
-  if (groupPrayCardList.length == 1) {
-    return (
-      <div className="flex flex-col justify-center items-center px-10 gap-4">
-        <p className="text-lg font-bold">아직 올라온 기도제목이 없어요 😭</p>
-        <div className="h-[300px] flex flex-col items-center">
-          <img className="h-full rounded-md" src="/images/KakaoShare.png" />
-        </div>
-        <div className="text-center">
-          <p className="text-sm text-gray-500">
-            친구들과 함께 오늘의 기도를 시작해 보아요:)
-          </p>
-        </div>
-        <KakaoShareButton
-          groupPageUrl={window.location.href}
-          id="paryTodayIntro"
-          message="카카오톡으로 초대하기"
-          eventOption={{ where: "PrayCardList" }}
-        ></KakaoShareButton>
-      </div>
-    );
-  }
-
-  const todayDt = getISOTodayDate();
   const filterdGroupPrayCardList = groupPrayCardList?.filter(
     (prayCard) =>
       prayCard.user_id !== currentUserId &&
       prayCard.pray?.filter((pray) => pray.created_at >= todayDt).length === 0
   );
 
-  return (
+  const emptyPrayCardList = (
+    <div className="flex flex-col justify-center items-center px-10 gap-4">
+      <p className="text-lg font-bold">아직 올라온 기도제목이 없어요 😭</p>
+      <div className="h-[300px] flex flex-col items-center">
+        <img
+          className="h-full rounded-md"
+          src="/images/KakaoShareMessage.png"
+        />
+      </div>
+      <div className="text-center">
+        <p className="text-sm text-gray-500">
+          친구들과 함께 오늘의 기도를 시작해 보아요:)
+        </p>
+      </div>
+      <KakaoShareButton
+        groupPageUrl={window.location.href}
+        id="paryTodayIntro"
+        message="카카오톡으로 초대하기"
+        eventOption={{ where: "PrayCardList" }}
+      ></KakaoShareButton>
+    </div>
+  );
+
+  const completedItem = (
+    <div className="flex flex-col gap-4 justify-center items-center min-h-[80vh] max-h-[80vh] pb-10">
+      <img
+        src={PrayTypeDatas["pray"].img}
+        alt={PrayTypeDatas["pray"].emoji}
+        className="w-16 h-16 opacity-100"
+      />
+      <h1 className="font-bold text-xl">오늘의 기도 완료!</h1>
+      <h3 className="text-gray-600">내일도 기도해 주실 거죠? 🤗</h3>
+      <div className="text-gray-400 text-center">
+        <h1>당신을 위해 기도한</h1>
+        <h1>친구들을 확인해 보아요</h1>
+      </div>
+      <MyMemberBtn />
+      <OpenContentDrawerBtn />
+    </div>
+  );
+
+  const todayPrayCardList = (
     <Carousel
       setApi={setPrayCardCarouselApi}
       opts={{
@@ -132,6 +148,29 @@ const TodayPrayCardList: React.FC<PrayCardListProps> = ({
         <CarouselItem className="basis-5/6"></CarouselItem>
       </CarouselContent>
     </Carousel>
+  );
+
+  return (
+    <Drawer
+      open={isOpenTodayPrayDrawer}
+      onOpenChange={setIsOpenTodayPrayDrawer}
+    >
+      <DrawerContent className="bg-mainBg max-w-[480px] mx-auto w-full pb-5">
+        <DrawerHeader className="p-2">
+          <DrawerTitle></DrawerTitle>
+          <DrawerDescription className="text-gray-300 text-xs text-center">
+            {!isPrayToday && filterdGroupPrayCardList.length > 0 ? (
+              "반응을 누르면 다음 기도로 넘어가요"
+            ) : (
+              <br />
+            )}
+          </DrawerDescription>
+        </DrawerHeader>
+        {filterdGroupPrayCardList.length === 0
+          ? emptyPrayCardList
+          : todayPrayCardList}
+      </DrawerContent>
+    </Drawer>
   );
 };
